Бампнуты зависимости; кнопка Help; изменено поведение Progress при DurationMs = 0

+++ NowPlaying.Wpf/Controls/PlayingTrack/PlayingTrackViewModel.cs
@@ -6,8 +6,8 @@ namespace NowPlaying.Wpf.Controls.PlayingTrack
     {
         [Reactive] public string Title { get; set; }
         [Reactive] public string Author { get; set; }
-        [Reactive] public int DurationMs { get; set; }
-        [Reactive] public int ProgressMs { get; set; }
+        [Reactive] public long DurationMs { get; set; }
+        [Reactive] public long ProgressMs { get; set; }
         [Reactive] public string CurrentProgress { get; set; }
         [Reactive] public string EstimatedProgress { get; set; }
     }
