+++ NowPlaying.Wpf/Controls/PlayingTrack/ProgressViewModel.cs
@@ -5,6 +5,6 @@ namespace NowPlaying.Wpf.Controls.PlayingTrack
     public class ProgressViewModel
     {
         // 0 <= Progress <= 100
-        [Reactive] public int Progress { get; set; }
+        [Reactive] public long Progress { get; set; }
     }
 }
